update short silence detection to return fraction of a pixel values, especially now that I have discovered frame boundaries appear to all be on 0.5 marks!

+++ .config/hammerspoon/config/macros/screenpal/py/short_silences.py
@@ -78,8 +78,9 @@ def detect_short_silences(use_file):
         "short_silences": [
             {
                 # divide by 2 for non-retina resolution
-                "x_start": int(x_start / 2),
-                "x_end": int(x_end / 2),
+                # preserve fraction!
+                "x_start": x_start / 2,
+                "x_end": x_end / 2,
             } for x_start, x_end in ranges \
                if x_start < x_end
         ],
